[luci/pass] Change QuantizeWeightsWithGPTQPass to QuantizeDequantizeWeightsWithGPTQPass
This commit renames QuantizeWeightsWithGPTQPass to QuantizeDequantizeWeightsWithGPTQPass and updates related references in the codebase.

ONE-DCO-1.0-Signed-off-by: y01000.you <y01000.you@samsung.com>
