+++ compiler/luci/pass/src/QuantizeDequantizeWeightsWithGPTQPass.test.cpp
@@ -1,4 +1,4 @@
-#include "luci/Pass/QuantizeWeightsWithGPTQPass.h"
+#include "luci/Pass/QuantizeDequantizeWeightsWithGPTQPass.h"
 #include <luci/IR/CircleNodes.h>
 
 #include <gtest/gtest.h>
@@ -73,22 +73,22 @@ struct QuantizeWeightsWithGPTQPassTest : public ::testing::Test
 
 TEST_F(QuantizeWeightsWithGPTQPassTest, name)
 {
-  luci::QuantizeWeightsWithGPTQPass pass(loco::DataType::FLOAT32, loco::DataType::U8,
-                                         luci::QuantizationGranularity::ChannelWise);
+  luci::QuantizeDequantizeWeightsWithGPTQPass pass(loco::DataType::FLOAT32, loco::DataType::U8,
+                                                   luci::QuantizationGranularity::ChannelWise);
   auto const name = pass.name();
   ASSERT_NE(nullptr, name);
 }
 
 TEST_F(QuantizeWeightsWithGPTQPassTest, name_ctx)
 {
-  auto ctx = std::make_unique<luci::QuantizeWeightsWithGPTQPass::Context>();
+  auto ctx = std::make_unique<luci::QuantizeDequantizeWeightsWithGPTQPass::Context>();
   {
     ctx->input_model_dtype = loco::DataType::FLOAT32;
     ctx->output_model_dtype = loco::DataType::U8;
     ctx->granularity = luci::QuantizationGranularity::ChannelWise;
   }
 
-  luci::QuantizeWeightsWithGPTQPass pass(std::move(ctx));
+  luci::QuantizeDequantizeWeightsWithGPTQPass pass(std::move(ctx));
   auto const name = pass.name();
   ASSERT_NE(nullptr, name);
 }
@@ -97,16 +97,16 @@ TEST_F(QuantizeWeightsWithGPTQPassTest, name_ctx)
 TEST_F(QuantizeWeightsWithGPTQPassTest, run_granularity_invalid_NEG)
 {
   auto invalid_granularity = static_cast<luci::QuantizationGranularity>(999);
-  luci::QuantizeWeightsWithGPTQPass pass(loco::DataType::FLOAT32, loco::DataType::U8,
-                                         invalid_granularity);
+  luci::QuantizeDequantizeWeightsWithGPTQPass pass(loco::DataType::FLOAT32, loco::DataType::U8,
+                                                   invalid_granularity);
   ASSERT_EXIT(pass.run(&_g), ::testing::KilledBySignal(SIGSEGV), ".*");
 }
 
 // Negative test: Unsupported output data type - FLOAT32
 TEST_F(QuantizeWeightsWithGPTQPassTest, run_output_f32_NEG)
 {
-  luci::QuantizeWeightsWithGPTQPass pass(loco::DataType::FLOAT32, loco::DataType::FLOAT32,
-                                         luci::QuantizationGranularity::ChannelWise);
+  luci::QuantizeDequantizeWeightsWithGPTQPass pass(loco::DataType::FLOAT32, loco::DataType::FLOAT32,
+                                                   luci::QuantizationGranularity::ChannelWise);
   // Since output type is FLOAT32, an exception is expected
   EXPECT_THROW(pass.run(&_g), std::runtime_error);
 }
@@ -115,14 +115,14 @@ TEST_F(QuantizeWeightsWithGPTQPassTest, run_output_f32_NEG)
 TEST_F(QuantizeWeightsWithGPTQPassTest, run_with_empty_hessian_map_NEG)
 {
   std::unordered_map<const luci::CircleNode *, std::vector<float>> hessian_map;
-  auto ctx = std::make_unique<luci::QuantizeWeightsWithGPTQPass::Context>();
+  auto ctx = std::make_unique<luci::QuantizeDequantizeWeightsWithGPTQPass::Context>();
   {
     ctx->input_model_dtype = loco::DataType::FLOAT32;
     ctx->output_model_dtype = loco::DataType::U8;
     ctx->granularity = luci::QuantizationGranularity::ChannelWise;
   }
 
-  luci::QuantizeWeightsWithGPTQPass pass(std::move(ctx), &hessian_map);
+  luci::QuantizeDequantizeWeightsWithGPTQPass pass(std::move(ctx), &hessian_map);
   // Expect no exception, pass should handle empty hessian map gracefully
   EXPECT_NO_THROW(pass.run(&_g));
 }
@@ -147,8 +147,8 @@ TEST_F(QuantizeWeightsWithGPTQPassTest, run_with_non_float_weights_NEG)
   // Set dtype to INT32
   weight->dtype(loco::DataType::S32);
 
-  luci::QuantizeWeightsWithGPTQPass pass(loco::DataType::FLOAT32, loco::DataType::U8,
-                                         luci::QuantizationGranularity::ChannelWise);
+  luci::QuantizeDequantizeWeightsWithGPTQPass pass(loco::DataType::FLOAT32, loco::DataType::U8,
+                                                   luci::QuantizationGranularity::ChannelWise);
   // The pass should skip this node without exception
   EXPECT_NO_THROW(pass.run(&_g));
 }
@@ -181,29 +181,29 @@ TEST_F(QuantizeWeightsWithGPTQPassTest, run_with_valid_hessian)
 
   hessian_map[conv] = hessian;
 
-  auto ctx = std::make_unique<luci::QuantizeWeightsWithGPTQPass::Context>();
+  auto ctx = std::make_unique<luci::QuantizeDequantizeWeightsWithGPTQPass::Context>();
   {
     ctx->input_model_dtype = loco::DataType::FLOAT32;
     ctx->output_model_dtype = loco::DataType::U8;
     ctx->granularity = luci::QuantizationGranularity::ChannelWise;
   }
 
-  luci::QuantizeWeightsWithGPTQPass pass(std::move(ctx), &hessian_map);
+  luci::QuantizeDequantizeWeightsWithGPTQPass pass(std::move(ctx), &hessian_map);
   EXPECT_NO_THROW(pass.run(&_g));
 }
 
 // Negative test: Input model data type is U8 (unsupported)
 TEST_F(QuantizeWeightsWithGPTQPassTest, run_input_U8_NEG)
 {
-  luci::QuantizeWeightsWithGPTQPass pass(loco::DataType::U8, loco::DataType::U8,
-                                         luci::QuantizationGranularity::ChannelWise);
+  luci::QuantizeDequantizeWeightsWithGPTQPass pass(loco::DataType::U8, loco::DataType::U8,
+                                                   luci::QuantizationGranularity::ChannelWise);
   EXPECT_THROW(pass.run(&_g), std::runtime_error);
 }
 
 // Negative test: Output model data type is S32 (unsupported)
 TEST_F(QuantizeWeightsWithGPTQPassTest, run_output_S32_NEG)
 {
-  luci::QuantizeWeightsWithGPTQPass pass(loco::DataType::FLOAT32, loco::DataType::S32,
-                                         luci::QuantizationGranularity::ChannelWise);
+  luci::QuantizeDequantizeWeightsWithGPTQPass pass(loco::DataType::FLOAT32, loco::DataType::S32,
+                                                   luci::QuantizationGranularity::ChannelWise);
   EXPECT_THROW(pass.run(&_g), std::runtime_error);
 }
